pyproject.toml
[project]
name = "python-for-ml"
version = "0.1.0"
description = "Python Programming For AI/ML"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
